get all doctors 2

+++ lib/authentication/data/response/user_response/user_response.dart
@@ -1,5 +1,6 @@
 // ignore_for_file: depend_on_referenced_packages
 
+import 'package:health_care/authentication/data/response/doctor_auth_response/doctor_auth_response.dart';
 import 'package:json_annotation/json_annotation.dart';
 
 import '../../../../core/response/base_response.dart';
@@ -15,7 +16,7 @@ class UserResponse {
   @JsonKey(name: "fees")
   int? fees;
   @JsonKey(name: "timePerPatient")
-  int? timePerPatient;
+  double? timePerPatient;
   @JsonKey(name: "ScheduleTiming")
   List<ScheduleTimingResponse>? scheduleTiming;
   @JsonKey(name: "patients")
